feat: Add SyncSketch addon documentation
This commit adds the documentation for the SyncSketch addon. It includes information on how to configure Ayon, set up secrets, and link projects between Ftrack and SyncSketch. Additionally, it provides instructions on configuring plugin settings and targeting reviewables to SyncSketch. The documentation also includes screenshots for reference.

The SyncSketch addon version is 0.1.1.

Co-authored-by: [Jakub Jezek] <jakub@ynput.io>